moved function files to common and updated notebooks imports.  added summary
