+++ common/analogy.py
